allow short names in remote interventions

+++ src/remote_switch.py
@@ -36,25 +36,76 @@ class remote_switch( Pyro.core.ObjBase ):
         self.halt_now = False
 
         self.locked = True
-
         self.owner = os.environ['USER']
 
-        self.process_task_names()
+        self.create_name_translation_table()
 
     def is_legal( self, user ):
         legal = True
         reasons = []
         if user != self.owner:
             legal = False
-            self.warning( "refusing task reset request (wrong owner)" )
-            reasons.append( "system owner: " + self.owner )
+            self.warning( "refusing remote request (wrong owner)" )
+            reasons.append( "wrong owner: " + self.owner )
+
         if self.locked:
             legal = False
-            self.warning( "REMOTE: refusing task reset request (locked)" )
+            self.warning( "refusing remote request (system locked)" )
             reasons.append( "SYSTEM LOCKED" )
 
         return ( legal, ', '.join( reasons ) )
 
+    def name_from_id( self, id ):
+        if '%' in id:
+            name, tag = id.split('%')
+        else:
+            name = id
+        return name
+
+    def create_name_translation_table( self ):
+        # create a mapping between (short and long) task names and their
+        # corresponding full names.
+
+        self.task_list = self.config.get( 'task_list' )
+        self.real_name = {}
+
+        clsmod = __import__( 'task_classes' )
+
+        for name in self.task_list:
+            cls = getattr( clsmod, name )
+            try:
+                short_name, = cls.short_name
+            except AttributeError:
+                # this task class has no short name defined
+                self.real_name[ name ] = name
+            else:
+                if short_name in self.real_name:
+                    raise SystemExit( "duplicate short task names detected" )
+
+                self.real_name[ short_name ] = name
+                self.real_name[ name ] = name
+
+    def task_type_exists( self, name_or_id ):
+        # does a task name or id match a known task type in this system?
+        name = name_or_id
+        if '%' in name_or_id:
+            name, tag = name.split('%' )
+        
+        if name in self.real_name:
+            return True
+        else:
+            return False
+
+    def translate( self, name_or_id ):
+        # translate possibly short names into proper task names
+        # if not self.task_type_exists( name_or_id ):
+        #   return name_or_id
+
+        if '%' in name_or_id:
+            name, tag = name_or_id.split('%' )
+            return self.real_name[ name ] + '%' + tag
+        else:
+            return self.real_name[ name ]
 
     def warning( self, msg ):
         print
@@ -71,10 +122,9 @@ class remote_switch( Pyro.core.ObjBase ):
 
     def unlock( self, user ):
         if user != self.owner:
-            return "ILLEGAL OPERATION: This system is owned by " + self.owner
+            return "This system is owned by " + self.owner
 
         if not self.locked:
-            self.warning( "REMOTE: system unlock requested (already unlocked)" )
             return "System already unlocked"
 
         else:
@@ -91,12 +141,16 @@ class remote_switch( Pyro.core.ObjBase ):
         # cause the task processing loop to be invoked
         self.warning( "nudging by remote request" )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
     def reset_to_waiting( self, task_id, user ):
         legal, reasons = self.is_legal( user )
         if not legal:
             return reasons
+
+        if not self.task_type_exists( task_id ):
+            return "No such task type: " + self.name_from_id( task_id )
+        task_id = self.translate( task_id )
 
         # reset a task to the waiting state
         self.warning( "REMOTE: reset to waiting: " + task_id )
@@ -105,21 +159,25 @@ class remote_switch( Pyro.core.ObjBase ):
             os.environ['FAILOUT_ID'] = ""
         self.pool.reset_task( task_id )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
     def reset_to_ready( self, task_id, user ):
         legal, reasons = self.is_legal( user )
         if not legal:
             return reasons
 
-       # reset a task to the ready state
+        if not self.task_type_exists( task_id ):
+            return "No such task type: " + self.name_from_id( task_id )
+        task_id = self.translate( task_id )
+
+        # reset a task to the ready state
         self.warning( "REMOTE: reset to ready: " + task_id )
         if task_id == self.failout_id:
             print "resetting failout on " + self.failout_id + " prior to reset"
             os.environ['FAILOUT_ID'] = ""
         self.pool.reset_task_to_ready( task_id )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
 
     def reset_to_finished( self, task_id, user ):
@@ -127,27 +185,41 @@ class remote_switch( Pyro.core.ObjBase ):
         if not legal:
             return reasons
 
-      # reset a task to the waiting finished
+        if not self.task_type_exists( task_id ):
+            return "No such task type: " + self.name_from_id( task_id )
+        task_id = self.translate( task_id )
+
+        # reset a task to the finished state
         self.warning( "REMOTE: reset to finished: " + task_id )
         self.pool.reset_task_to_finished( task_id )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
 
-    def insert( self, ins, user ):
+    def insert( self, ins_id, user ):
         legal, reasons = self.is_legal( user )
         if not legal:
             return reasons
 
+        ins_name = self.name_from_id( ins_id )
+
+        if not self.task_type_exists( ins_name ) and \
+                ins_name not in self.config.get( 'task_groups' ):
+            return "No such task type or group: " + ins_name
+
+        ins = ins_id
+        if self.task_type_exists( ins_name ):
+            ins = self.translate( ins_id )
+
         # insert a new task or task group into the system
-        self.warning( "REMOTE: task insertion: " + ins )
+        self.warning( "REMOTE: task or group insertion: " + ins )
         if ins == self.failout_id:
             # TO DO: DOES EQUALITY TEST FAIL IF INS IS A GROUP?
             print "resetting failout on " + self.failout_id + " prior to insertion"
             os.environ['FAILOUT_ID'] = ""
         self.pool.insertion( ins )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
 
     def hold( self, user ):
@@ -159,7 +231,7 @@ class remote_switch( Pyro.core.ObjBase ):
         self.pool.set_system_hold()
         # process, to update state summary
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
 
     def resume( self, user ):
@@ -171,7 +243,7 @@ class remote_switch( Pyro.core.ObjBase ):
         self.pool.unset_system_hold()
         # process, to update state summary
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
 
     def set_stop_time( self, ctime, user ):
@@ -183,7 +255,7 @@ class remote_switch( Pyro.core.ObjBase ):
         self.pool.set_stop_time( ctime )
         # process, to update state summary
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
     def set_hold_time( self, ctime, user ):
         legal, reasons = self.is_legal( user )
@@ -194,8 +266,7 @@ class remote_switch( Pyro.core.ObjBase ):
         self.pool.set_system_hold( ctime )
         # process, to update state summary
         self.process_tasks = True
-        return "Done"
-
+        return "OK"
 
     def shutdown( self, user ):
         legal, reasons = self.is_legal( user )
@@ -207,8 +278,7 @@ class remote_switch( Pyro.core.ObjBase ):
         self.halt = True
         # process, to update state summary
         self.process_tasks = True
-        return "Done"
-
+        return "OK"
 
     def shutdown_now( self, user ):
         legal, reasons = self.is_legal( user )
@@ -220,9 +290,7 @@ class remote_switch( Pyro.core.ObjBase ):
         self.halt_now = True
         # process, to update state summary
         self.process_tasks = True
-        return "Done"
-
-
+        return "OK"
 
     def get_config( self, item ):
         self.warning( "REMOTE: config item " + item )
@@ -243,26 +311,34 @@ class remote_switch( Pyro.core.ObjBase ):
     def get_task_info( self, task_names ):
         self.warning( "REMOTE: task info: " + ','.join(task_names ))
         info = {}
-        for n in task_names:
-            try:
-                descr = eval( 'task_classes.' + n + '.describe()' )
-            except AttributeError:
-                info[ n ] = ['ERROR: No Such Task Class']
+        for name in task_names:
+            if self.task_type_exists( name ):
+                info[ name ] = eval( 'task_classes.' + self.real_name[ name ] + '.describe()' )
             else:
-                info[ n ] = descr
-
+                info[ name ] = ['ERROR: no such task type']
         return info
 
-    def get_task_requisites( self, task_ids ):
-        self.warning( "REMOTE: requisite request")
+    def get_task_requisites( self, in_ids ):
+        self.warning( "REMOTE: task requisite dump request")
+
+        in_ids_real = {}
+        in_ids_back = {}
+        for in_id in in_ids:
+            if not self.task_type_exists( in_id ):
+                continue
+            real_id = self.translate( in_id )
+            in_ids_real[ in_id ] = real_id
+            in_ids_back[ real_id ] = in_id
 
         dump = {}
         found = False
         for task in self.tasks:
+            # loop through the system task list
             id = task.id
-            if id in task_ids:
-                found = True
 
+            if id in in_ids_back:
+
+                found = True
                 extra_info = {}
 
                 # extra info for contact tasks
@@ -280,10 +356,10 @@ class remote_switch( Pyro.core.ObjBase ):
                     # not a catchup_contact task
                     pass
 
-                dump[ id ] = [ task.prerequisites.dump(), task.outputs.dump(), extra_info ]
+                dump[ in_ids_back[ id ] ] = [ task.prerequisites.dump(), task.outputs.dump(), extra_info ]
 
         if not found:
-            self.warning( 'No tasks found for the requisite dump request' )
+            self.warning( 'No tasks found to dump' )
  
         return dump
     
@@ -292,21 +368,28 @@ class remote_switch( Pyro.core.ObjBase ):
         if not legal:
             return reasons
 
+        if not self.task_type_exists( task_id ):
+            return "No such task type: " + self.name_from_id( task_id )
+        task_id = self.translate( task_id )
+
         self.warning( "REMOTE: purge " + task_id + ' to ' + stop )
         self.pool.purge( task_id, stop )
         self.process_tasks = True
-        return "Done"
-
+        return "OK"
 
     def die( self, task_id, user ):
         legal, reasons = self.is_legal( user )
         if not legal:
             return reasons
 
+        if not self.task_type_exists( task_id ):
+            return "No such task type: " + self.name_from_id( task_id )
+        task_id = self.translate( task_id )
+
         self.warning( "REMOTE: die: " + task_id )
         self.pool.kill( [ task_id ] )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
     def die_cycle( self, cycle, user ):
         legal, reasons = self.is_legal( user )
@@ -316,17 +399,21 @@ class remote_switch( Pyro.core.ObjBase ):
         self.warning( "REMOTE: kill cycle: " + cycle )
         self.pool.kill_cycle( cycle )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
     def spawn_and_die( self, task_id, user ):
         legal, reasons = self.is_legal( user )
         if not legal:
             return reasons
 
+        if not self.task_type_exists( task_id ):
+            return "No such task type: " + self.name_from_id( task_id )
+        task_id = self.translate( task_id )
+
         self.warning( "REMOTE: spawn and die: " + task_id )
         self.pool.spawn_and_die( [ task_id ] )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
     def spawn_and_die_cycle( self, cycle, user ):
         legal, reasons = self.is_legal( user )
@@ -336,7 +423,7 @@ class remote_switch( Pyro.core.ObjBase ):
         self.warning( "REMOTE: spawn and die cycle: " + cycle )
         self.pool.spawn_and_die_cycle( cycle )
         self.process_tasks = True
-        return "Done"
+        return "OK"
 
     def set_verbosity( self, level, user ):
         legal, reasons = self.is_legal( user )
